feat : 하이브 프리팹 완

+++ Assets/Scripts/Entities/Enemy/Stage1/Hive/HiveSO.cs
@@ -8,7 +8,7 @@ namespace ToB
     {
         [field:SerializeField] public float HP { get; private set; }
         [field:SerializeField] public float FlyRegenTimeInterval { get; private set; }
-        [field:SerializeField] public float FlyRegenAmount { get; private set; }
+        [field:SerializeField] public int FlyRegenAmount { get; private set; }
         [field:SerializeField] public float PatrolRange { get; private set; }
         [field:SerializeField] public float ChaseRange { get; private set; }
